Change toPollingObservable to toPollingRepository
In the new implementation of watch in sbt, I replace the global file
system repository with a local repository that exists only for the
duration of the watch process. This allows me to intercept inputs during
task evaluation that may have been hidden from the dependency graph
because they were used within a dynamic task. I then register those
intercepted inputs with the file tree repository. In the
HybridPollingFileTreeRepository, we do not want to constantly run the
polling process in the background if we are not in a watch. The way to
make watching work, then, is to create a new FileTreeRepository during
the watch process that persistently polls until the watch ends.