Optimize delete method
The previous implementation of delete did a lot of unnecessary io
because it had to check whether or not the file was a directory if
deleting the file failed. The view.list method returns Seq[TypedPath]
for which the isDirectory property is memoized (and is calculated
without an additional stat).