Bump swoval version
I discovered a nasty performance bug that impacted windows and linux in
swoval. Whenever the background watch process for the global file
repository detected that a directory was modified, it would rescan the
entire directory to determine what had changed. This could cause a huge
amount of io to happen if there were a lot of files in the modified
directory. It also was generally unnecessary because the directory
should have been updated via incremental file updates. This change does
make it slightly more likely that the cache will get out of sync with
the file system, but the cache is easily flushed and users can disable
it if they have any issues anyway.