Convert glob base path to absolute path
If one does not convert the glob path to an absolute path, then the
filter that checks that the path starts with the base does not work. In
general, we want to use absolute paths as much as possible even though
there may be a memory overhead to doing so.

This was needed because the project/extra scripted test started failing
once I updated swoval to correctly handle relative paths. There was a
line like `(file("lib_managed") ** "*.jar").get` that was returning no
results even though there were jars in that directory. This was because
of the aforementioned filtering issue relative to the base directory.