Refactor all of the glob/watch related code
I came up with the FileTreeRepository interface roughly a year ago when
I was working on the close watch plugin. Since then, my understanding of
the problem has improved and I see that the interfaces that I defined
were not quite as clean as they could have been. The core problem that
the file cache tries to solve is that both listing _AND_ filtering of
files can be slow in sbt. Since the beginning, sbt has primarily used
java.io.File based apis for io. The problem with this is that if a user
wants to filter by file attributes, this will almost always lead to
extra io. That extra io will in turn lead to increased latency before
sbt can start compiling.  If we continue to use file based apis, this
will always be a problem.

I already started tackling this problem by introducing globs. The Glob
spec makes it possible to efficiently retrieve the source files.
Moreover in the current sbt version, I employ some dirty tricks to
return java.io.File instances that have their isRegularFile and
isDirectory methods overridden so that they use a memoized value
computed while we recursively fetch the sources. I realized, however,
that this was an ugly hack and still required the task author to think
about how downstream consumers of the source file might use them.

To improve the situation, I rethought the interface for fetching files.
Instead of effectively having a function from Glob => Seq[File], I want
to instead have a function from
(Glob, AttributeFilter) => Seq[(Path, Attributes)]
Path is better than file because it gives us the option of swapping out
the file system. Moreover, by returning the Attributes, we can
explicitly filter paths based on their attributes without having to do
any io.

This new api allowed me to simplify Glob so that instead of having a
filter from TypedPath => Boolean, it now just has a filter from
Path => Boolean. Most of the path methods do not do any io except for
Path.toRealPath(). This means that the Glob filter will almost always be
fast since it's likely just doing a file name check. Part of the
motivation for this change is that TypedPath was an awkward interface
since it combined file meta data (i.e. the path) with the file
attributes. To replace the functionality, I introduce two traits
SimpleFileAttributes and CustomFileAttributes. SimpeFileAttributes is
analogous to TypedPath without the getPath method while
CustomFileAttributes is a replacement for FileTreeDataView.Entry.

The reason for adding the CustomFileAttributes class and having it
implement the BasicFileAttributes interface is because it gives the
option of using it as a valid file attribute type in a custom file
system. The BasicFileAttributes interface partially mirrors the posix
stat struct which means to implement it, you pretty much need to stat
the file. We don't, have to eagerly implement it, however. The default
implementation implements SimpleFileAttributes (which are basically free
as part of the swoval directory listing api) and delegates the other
methods to a lazily computed default BasicFileAttributes instance.

With respect to listing files, I simplified the FileTreeView api.
Instead of returning a Seq[TypedPath], it returns a Seq[T] for some
generic type T. This allowed me to remove the largely redundant
FileTreeDataView interface. In practice, we mainly only care about
FileTreeView[(Path, SimpleFileAttributes)]
FileTreeView[(Path, CustomFileAttributes[T])]. The former maps closely
to the old default FileTreeView while the latter is similar to the old
FileTreeDataView.

I simplified the Observer interface and removed the redunant
CacheObserver interface. Instead of having the CacheObserver interface,
I make the various file tree repository instance implement
Observable[FileEvent[T]] where T is the cache value. Implementing an
Observer[FileEvent[T]] is very similar to implementing a CacheObserver
and we remove a redundant interface.

I formalized and improved the Registerable interface. When one
registers a glob with a FileTreeRepository, they probably want to
receive events specifically for that glob. I make that possible by
having register return Either[IOExceptin, Observable[T]] instead of
Either[IOException, Boolean]. This will be helpful in the watch refactor
where I dynamically register globs during task evaluation and want to
monitor them for watch events. Previously, I had to manage the filtering
to exclude, say, jar files that are were also registered with the
repository. With the new register interface, I can add a simple api in
sbt to mark which globs you want to watch for events in the task
definition. When we list a marked glob, we'll register it with the file
repository and add it to the other observables that we're already
monitoring.

Even though a lot of code had to change to implement the refactored
interfaces, most of the code changes were very obvious since there was a
pretty clear 1:1  mapping between the new and old interfaces. For the
most part, the tests would work with minimal tuning once they compiled.

Finally, I tried to ensure that all of the apis related to file
repository are marked private[sbt]. I'm still not sure if I've gotten
these interfaces right and I want the freedom to change them without
breaking binary compatibility. In subsequent commits (or a new PR), I
will move more of the implementation classes into sbt.internal.io.