Rewrite PollingWatchService
I was seeing occasional failures of the PollingWatchService test both in
CI and on my local machine. To try and address this, I rewrite the
service so that it does not use a polling thread. Instead, each key
maintains a FileCache and when we call WatchService.poll, instead of
reading from a queue filled by a thread, we iterate through the watch
keys and poll them on demand. Since FileCache already has logic for
refreshing the path and generating events for all of the changed paths,
we don't need to re-implement that logic. This makes the implementation
simpler. Avoiding the background thread makes it easier to reason about
what is going on/reduces race conditions.