Improve SourceModificationWatchSpec
Previously the watchTest method was specified so that it succeeded
either if any file triggered or if no files triggered. There were two
issues with this:
1) We weren't actually validating if the expected file triggered
2) Waiting for a timeout both adds a lot of time to the test runs and
   can lead to flakiness

To fix this, I reworked watchTest so that the caller must provided a
filter of file events that they are interested in as well as a
validation function from Seq[FileEvent[_]] => Boolean to check that the
right type of events were produced.

To verify that certain events do not trigger, I just rework the tests so
that I modify the ignored paths and then modify a monitored path and
assert that there were no events generated for the ignored path. This
makes the tests more than 2x faster and makes it much more explicit what
is being tested which is a huge plus.