Change depth parameter to range in Glob
A user may want to specify a glob (in shell syntax) like: foo/*/*.md
This was not really possible with the previous api because only the
maximum depth of the glob was configurable. To remedy this, I changed
the depth `Int` parameter to the range `(Int, Int)` parameter. I also
changed the indexing scheme. The depth of a path with respect to a glob
is now calculated by computing the number of path names when the path is
relativized with respect to the glob base path. In this scheme, a range
of (0, 0) means only include the base path. A range of (0, 1) means
include the base path and its immediate children. A range of
(1, Int.MaxValue) means include all recursive descendants of the base
path, but not the base path itself.