Remove LazyFileAttributes
Since I needed to evaluate the BasicFileAttributes in order to evaluate
the SimpleFileAttributes I thought it made sense to have
CustomFileAttributes extend BasicFileAttributes. This was actually
problematic though because when creating SimpleFileAttributes instances
when listing a directory using the glob apis, we don't actually need to
compute the BasicFileAttributes to create an instance of
SimpleFileAttributes. Evaluating the BasicFileAttributes requires
stating the file, which is expensive so we lose the performance benefit
of using swoval in the first place vs. the nio apis. To work around
this, I'd made the BasicFileAttributes evaluation lazy conditional on
whether or not the user ever called one of the BasicFileAttributes
methods not implemented by SimpleFileAttributes. It's a better idea to
just not extend BasicFileAttributes at all since we don't need them in
the sbt use case.