Repackage new classes
It makes sense to isolate the new io apis in a new package. Lumping
everything into io made it difficult to be opinionated about which
api to use. For example, if sbt.io.syntax has methods for converting
both File and Path to a Glob, we are implicitly saying that there is
no preference for using Path vs File. Moving forward to sbt 2, we
would like to remove the File based apis entirely so isolating the
apis that we wish to preserve in a different package makes sense.

For whatever reason, git had trouble tracking the moved files in this
commit. There is very little in the way of functional changes. The one
important addition is that I added the PathFilter class to nio that
provides a subset of the filters provided by NameFilter only the apis
takes `Path`s as input instead of `File`s. I also added better
conversions from `FileFilter` to `PathFilter` so that Glob equality
works better even when Globs are specified using `FileFilter`, which is
the case in e.g.  `fileInputs += baseDirectory.value ** "*.txt".