Simplify interfaces
Previously the FileTreeRepository implementations maintained a cache of
(FileAttributes, T) for some generic type T. This was unneeded
complexity that could be managed with an external map from Path -> T.
The trouble with including the arbitrary data was that we needed to
compute it whenever we detected an update and when we prefilled the
cache. That could be expensive. In particular this led to slowdowns in
sbt because we ended up computing the files stamps (either hashes or
last modified times) for effectively all of the files in in the project
directory. The FileAttributes are computed for free when prefilling the
cache. We still have to stat the file to recompute the FileAttributes
when an event is detected, but that is significantly less expensive than
computing the file hash.

It also didn't really make sense that FileTreeView implemented
AutoCloseable so I dropped it from the trait definition. I additionally
removed the filter parameter from the list method since it wasn't used
anywhere.