Further simplify FileTreeView interface
I have become a big fan of the simple interface + ops class method. The
simplest interface for FileTreeView is Path => Seq[T]. It is very easy
to implement this interface but implementing Glob => Seq[T] is not
trivial. The benefit to this change is that by implementing Glob =>
Seq[T] in the ops class, it is dead simple to change the implementation
of the FileTreeView for a particular platform. Moreover because the Glob
=> Seq[T] (and Traversable[Glob] => Seq[T]) methods are defined in the
implicit ops class, we can can change those interfaces at the source
level without breaking binary compatibility.