Add api for listing a collection of globs
It is often the case that a user might want to list files from a
collection of globs. In sbt, this occurs when we list the sources for
the compile task because there are a number of source directories
specified (e.g. managed and unmanaged sources are in different
directories). It is not completely straightforward to list a collection
of globs because there may be overlapping paths. Ideally we would not
re-list each directory that is covered by multiple globs. To handle
this, I add the Glob.all method which maintains a list of visited paths
to ensure that we do not list the same path multiple times. Each path in
the result set is filtered against all of the globs in the collection to
ensure that we capture all of the relevant paths in each directory.