Add syntactic sugar for constructing globs
This commit adds a new dsl for constructing Globs from Path, String and
Glob instances. The rules are fairly restrictive. The Glob must start
with a path followed by an arbitrary number of '*'s separated by '/'
followed by an optional '**' followed by an optional path name filter.
For example, /foo/bar/*/*/*.scala and foo/bar/*/**/*.scala are valid
globs but /foo/bar/**/*/*.scala is not valid. Neither is
/foo/bar/**/src/**/*.scala valid. This is because the glob data
structure itself is not recursive.

The only name filters supported by the dsl are extension filters, exact
name filters and simple glob filters of the form [prefix]?*[suffix]?
when both prefix and suffix are empty, this is just equivalent to the
AllPass filter.

Globs are intended to map closely to Posix glob
zsh or bash. With the old PathFinder dsl, we provided syntactic sugar
for finding files in a directory that looked like, e.g.
baseDirectory.value / "src" / "main" / "scala" ** "*.scala"

This syntax was a little weird, especially when no filtering was
performed, e.g.
baseDirectory.value ** AllPassFilter
Since globs are part of an overall shift to using nio based apis, I
think it makes sense to update the api to make it more intuitive. To
this end, I added syntactic helpers that allow us to rewrite the first
example above as
baseDirectory.value.toPath / "src" / "main" / "scala" / "**" / "*.scala"
or
s"${baseDirectory.value}/src/main/scala/**/*.scala"
and the second example as
baseDirectory.value.toPath / "**"
or
s"${baseDirectory.value}/"**"}"

The GlobSyntaxSpec demonstrates the various way that users can create
globs.

I made sure that the old file based PathFinder syntax can be used
interchangeably. In many cases the syntax is very similar. The main
difference is that the '/' operator when applied to a Path or Glob
always returns a Glob. This contrasts with the File ops class for which
the '/' operator returns another `File`. However, the user can always
do, e.g.
(baseDirectory.value.toPath / "foo" / "bar").base
if they want the path represented by the glob created using the slash
syntax.

It is important that the toString method of Glob works correctly so that
it can be used to extend globs with the /(path: String) method in
GlobOps. To that end, there are a number of tests for toString. As a
bonus, this makes the output of the fileInputs task in sbt very easy to
read.