Improve glob parsing
In code review, @eed3si9n pointed out that globs did not actually work
exactly like shell globs. I for the most part remedied this by using the
matcher provided by java.nio.file.FileSystem.getPathMatcher. The logic
is a bit complicated because I still need to extract the range for the
glob even when using more exotic globs. This is actually not too bad,
however, because every slash in the relative pathname increments the
range by 1. The exception is when there is a recursive path specified by
"**". In this case, the maximum range parameter is set to Int.MaxValue.

There are almost certainly some weird cases in which the glob specified
by the user does not exactly match the glob generated by the parser, but
I can fix those in patch releases. This parser works fine for all of the
globs present in the sbt defaults.