Add syntax
This adds syntax that allows components to be added to Path, similar to
what already exists for file, so that baseDirectory.value.toPath / "foo"
returns a Path that is the child "foo" of the base directory. It also
adds a method for lifting a path to a Pattern glob. Although there is an
implicit conversion from String => Relative, scala won't perform nested
implicit conversions so the right hand side of `/` must be an instance
of Relative.

While there is some controversy around symbolic operators. I strongly
feel that the readability and convenience of
baseDirectory.value.toPath / "src" / ** / "*.scala"
compared to
baseDirectory.value.toPath / "src" / Recursive / "*.scala"
is worth the cognitive load of a value with a symbolic name.