Throttle polling in PollingWatchService
I realized that I was incorrectly allowing the PollingWatchKey to poll
the file system at any time. What we actually want is for at least the
PollingWatchService's delay parameter to elapse between polls. This
limits the polling watch service from hammering the disk

Bonus: imporve the toString of PollingWatchEvent for debugging.