Don't return empty files in parseClasspath
Zinc will pass in an empty string into this method if a system property
isn't set. It doesn't make sense to handle the empty string as a
relative file. I noticed this because sbt now prints a warning if one
tries to list a relative glob (which is generated by the pathfinder dsl
in zinc).