Add translation api for io <-> nio
In sbt, there are a number of tasks that define an
includeFilter and excludeFilter and use those filters when listing the
source or resource files for a task. These do not exactly map to the new
fileInputs glob based apis. To facilitate translating the old apis to
the new apis, I introduce the FullFileGlob class which is effectively
like an sbt.internal.io.Source with just one filter. I then add an
internal object Globs, that provides a factory apply method that takes a
base path, a recursive flag and a FileFilter and returns a Glob. In many
cases, we can convert the FileFilter to an equivalent Matcher. In those
where we can't, we return a FullFileGlob.

I add a number of tests that the conversions work as expected.