Fix bug in file events api for overlapping directory names
There is a bug in the osx file events api where if you register
/foo/bar and /foo/barfoo, then you will only get events with /foo/bar. I
verified that the bug is happening in the jni layer of swoval so we have
to work around it since it is an os issue. The workaround is to register
the common parent with the file events api and then you get results for
both.

Fixes https://github.com/sbt/sbt/issues/4603