Fix glob conversion from certain file filters
Issue https://github.com/sbt/sbt/issues/4687 reports that sbt 1.3.0-RC1
does not work correctly with the sbt header plugin. This is because the
plugin uses the Defaults.collectFiles method that was changed to use
globs. There was an issue with the base path + file filter conversion
where if the filter was, say, AllPassFilter && !HiddenFileFilter, this
actually got translated to $baseDiretory/**/<null> (where <null>
indicates the NoPath filter). The primary cause of this issue was an
incorrect conversion from HiddenFileFilter to AnyPath. Hidden files
are automatically excluded by globs so the HiddenFileFilter is actually
equivalent to the NoPath filter. I also added the NotHiddenFileFilter
object to be the inverse of the HiddenFileFilter because
-HiddenFileFilter comes up in a lot of places so I think it makes sense
to handle it directly.