Fix extension filter
I noticed that sbt 1.3.0 had problems with the twirl template compiler
and I tracked it down to the template compiler calling effectively
sourceDirectories ** "*.template.scala". This created an instance of
ExtensionFilter with the extension "template.scala". The problem was
that ExtensionFilter had an implicit assumption that the extension
didn't have any additional dots in it. This allowed it to consolidate
"*.java" || "*.scala" into ExtensionFilter("java", "scala"). This
optimization does not really work for more complicated extensions so I
added a check that the extension doesn't have any dots in it.