Fix FileAttributes for symbolic links
The semantics of sbt.nio.FileAttributes were not well specified before
and not what I intended.

When the followLinks parameter is set to true, then if a file is a
symbolic link, then the returned FileAttributes should always have
isSymbolicLink set to true. If the link is broken, then all of the other
parameters should be false. Otherwise, the parameters should be set
according to the file type.

When the followLinks parameter is set to false, then if the file is a
symbolic link, then isSymbolicLink should be true and all of the other
parameters should be false.

Unfortunately, there is no way to tell if the FileAttributes for a file
that has isSymbolicLink = true but all other parameters false is a
broken link or if FileAttributes was invoked with followLinks set to
false. For this reason, I made the FileAttributes.apply method that
takes the parameter private[sbt]. In sbt, we always seem to want
followLinks to be true anyway.

This also fixes a bug that a NoSuchFieldException was caught where I
meant to catch a NoSuchFileException. That was actually the motivation
for this change.