Include hidden files by default in globs
After some careful consideration, I think that it makes sense to include
hidden files in glob search queries. It makes it more straightforward to
document and less unpredictable. In the common case of excluding dot
files, it is easy enough to write
`baseDirectory.value / ** / "[!.]*.scala"`
or just add a manual hidden file filter of the results.