Optimize FileTreeView for file globs
I realized that if a user specifies a number of files as the fileInputs,
for example foo/Foo.scala and foo/Bar.scala instead of foo/*.scala, that
we'd do a lot of unnecessary io. To improve this, I reworked the
FileTreeView.iterator method to be smarter when it is listing a file
rather than a directory. When a file is detected, it will now list the
parent directory and fill an in memory cache. When the next file is
encountered, it will check if the cache has been filled for the path's
parent. Because we list the globs sorted by base path, we only need to
keep around the cache entry for the current file's parent path.

To make things more efficient, I changed the FileTreeView
implementations to throw IOExceptions instead of swalling them and
returning an empty Seq. This allows us to catch, NotDirectoryException
in the iterator method to know when we need to list the parent directory
of a file rather than reading the attributes directly. The new exception
behavior is only relevant if users directly call
FileTreeView.list(path: Path). Most of the time, they will be calling
one of the list methods that takes Glob or Traversable[Glob] as input.