Add PathFilter
In sbt 1.3.0, there are a few places where when using the Glob apis, we
always exclude hidden files. This should be configurable by the user
through a filter setting like they presently can with `includeFilter`
and `excludeFilter`. To this end, I introduce `PathFilter` which is a
SAM from `(Path, FileAttributes) => Boolean`. The idea here is that
`Glob` handles file name filtering while `PathFilter` handles other
properties of the file like whether it's a directory or hidden. Because
FileAttributes only provide a limited set of attributes, it is necessary
for the function to take a `Path` as well as a `FileAttributes` input.

I added some basic filters like HiddenFilter and DirectoryFilter. Using
extension methods, filters can be negated with `!` and combined with
`&&` and `||`.

Before the user would have had to do something like:

```
val scalaFiles =
  fileTreeView.list(baseDirectory.value.toGlob / "src" / ** / "*.scala").filter {
    case (p, _) => !Files.isHidden(p)
  }
```

Now they can write:
```
val scalaFiles =
  fileTreeView.list(baseDirectory.value.toGlob / "src" / ** / "*.scala", !HiddenFilter)
```
The new version is also somewhat better optimized because the excluded
files are never added to the result set.