Optimize FileTreeView.iterator for exact paths
It can get slow to check if each path in the results is matched by a
glob if there are many exact path globs. As an option, I partition the
input globs into exact file globs and other kinds of globs. The exact
file globs that have the same parent can be aggregated into a single
glob which can significantly reduce the number of calls to
`Glob.matches` that we have to make. In the "handle many exact file
globs" test, I saw a speedup from about 20ms to 1ms.

This is something of an edge case, but I thought of it when thinking
about writing an incremental task with Make semantics. Instead of
specifying wildcard globs, the user might want to manually enumerate
each file that the task operates on. It is somewhat unlikely that the
user would often specify enough files for this optimization but I think
it's still worth having since it doesn't change any of the core logic of
listing a sequence of globs so the maintenance cost is reasonably low.