Remove direct PathFinder inheritance from Glob
Even though Glob can be thought of as a specialization of PathFilter, it
will give us more flexibility if it does not directly inherit from
PathFilter. Instead, I add an implicit conversion from Glob to
PathFilter and add the PathFilter extension methods to GlobOps.

This also leaves open the door to better optimize the `&&` and `||`
methods for `Glob` since we can implement `PathFilterExtensions`
differently in `Glob.PathFilterOps` compared to `PathFilter.Ops`.