Add retries to FileTreeView.list
In 52cb29be7f911d73e501f52b76659d8efd6ad49c, we stopped catching all
exceptions in FileTreeView.list. This exposed an issue on windows where
sometimes an AccessDeniedException was thrown but what was quietly
ignored. To work around this, add retries when neither
NotDirectoryException nor NoSuchFileException are thrown. This is
really just to handle AccessDeniedExceptions on Windows.