Drop all errors in IO.delete
There is an expectation that IO.delete never throws an IOException. On
windows, scripted tests sometimes try to delete directories that are
still in use. If we don't catch all IOExceptions here, scripted tests
will fail on Windows. We could catch the exception at the call site, but
it seems likely that there are builds/consumers of sbt that expect this
method to not throw.