Don't require absolute paths in toGlob calls
While working on a plugin, I discovered that calling toGlob on a
relative path would cause an exception to be thrown. The workaround was
to call Glob(path.toString). There are valid uses cases for converting
paths to relative globs though, so I think it should be supported.

I added special handling for empty paths in the conversion because if
paths are programatically generated they may be empty. For example, in
`val glob = foo.resolve(bar).toGlob / **` , then the base path name
might be empty if `foo == bar`. This would cause the glob to pretty much
never work because paths rarely contain an empty path.