Apply DeMorgan's rule to not filters
I noticed that while the unary `!` operator worked correctly in sbt, the
toString looked wrong because NotPathFilter toString was defined as:
`override def toString: String = !filter`
When the filter was an || filter, this turned into something like
!DirectoryFilter || HiddenFileFilter
which was confusing because it actually represented
!(DirectoryFilter || HiddenFileFilter)

When I thought about it, I realized it would be nicer to just apply
DeMorgan's rules in the NotFilter. The filter then becomes:
!DirectoryFilter && !HiddenFileFilter
which I think is easier to read anyway.