Fix stack overflow in Glob ordering
This issue was reported in https://github.com/sbt/sbt/issues/4970. The
isssu was that the glob ordering had an infinite loop in it when
a Root glob was compared to a FullFileGlob (which are created to convert
directory + fileFilter to Glob for legacy support). The reason this
hadn't been seen is because the play plugin is one of the few places
that creates sbt.internal.io.Source from single paths, which is from where
the Root globs (which are rarely a part of the watch globs) are coming.
This seems to be a side effect of play putting assets in the source
directory:
watchSources ++= {
    ((sourceDirectory in Compile).value ** "*" --- (sourceDirectory in Assets).value ** "*").get
},
is the setting that triggered the issue.