Fix and comprehensively test glob ordering
The Ordering[Matcher] implementation violated the general ordering
contract. https://github.com/sbt/sbt/issues/4970 was still present
with sbt 1.3.0-RC5. This commit fixes the contract of Ordering[Matcher]
so that it should be sound for all Matcher instances.

Using jacoco, I was able to identify which branches of
the matchers were not being hit. After this change, all of the branches
were hit. While I'm not sure what was specifically causing the issue in
sbt/4970, I suspect that somewhere was an AndFilter, OrFilter or
NotFilter. The conversion in Globs from AndFilter, NotFilter and OrFilter to
Matcher created AndMatcher, NotMatcher and OrMatcher. The contract
between these matchers was not sound in the Ordering[Matcher]
implementation. Instead of trying to sort these, I just throw up my
hands and say that they are uncomparable.

After this change, I have 100% coverage of all of the branches in the
ordering implementations. Note that the ordering is unspecified between
AndMatcher, NotMatcher and OrMatcher. This means that, for example,
{ **/!foo, **/(foo && bar) } and { **/(foo && bar), **/!foo }
are both possible orderings depending on the initial ordering of these
two globs. However,
{ **/bar/!foo, **/foo/!bar }
is the only possible ordering for those two globs. In other words,
ordering is preserved up until two components differ only by an
(And|Not|Or)Matcher.