Close stream used in FileTreeView.nio
We created a stream without closing it. This came up because I was
reading the release notes for metals 0.7.2 and they mentioned fixing a
similar bug in the release notes.

In general, the nio FileTreeView should not really be in use in sbt
1.3.0. It would only be used on 32 bit platforms running sbt. Because
FileTreeView is a new api, and I rolled back the default path finder
implementations, none of the downstream consumers should be affected.