Wakeup FileEventMonitor when quarantine period ends
When the anti-entropy FileEventMonitor detects a deletion, it
quarantines the deletion for some period of time because some editors
implement save with a delete + rename. If we trigger on the initial
delete, we may end up for example starting compile while there is a
pending file write. This can be very painful if zinc ends up
invalidating all of the code in the deleted file only for the file to
reappear immediately thereafter. The quarantine implementation did not,
however, handle the case in which the polling period was longer than the
quarantine period. When this is the case, we only want to poll on the
original file event monitor no longer than the amount of time we need to
wait for the first quarantined deletion to be released from quarantine.

Prior to this change, if the anti-entropy FileEventMonitor polling
period was infinite, then a single file deletion would not be reported by
the anti-entropy FileEventMonitor until a different file change was
detected. We didn't notice this before because sbt was only polling with
small finite durations. sbt would use a fair amount of cpu while idle
file watching (~2% on my computer*) because it kept waking itself up to
check for new events. With an infinite poll the utilization drops to 0%.

* Measured by `top`. I do see something like 0-0.3% when I look at the
java process in the mac activity monitor even with with infinite poll.