Relax anti-entropy quarantine constraint
In sbt scripted tests, it was possible for files to trigger a few times
with the same timestamp which would cause the anti entropy file event
monitor to drop the event even if the anti-entropy period was set to
zero. This could be worked around with a negative anti-entropy period,
but this seems like a better fix.