Add sleep duration parameter to retry
I noticed that the RetrySpec was taking about 10 seconds to run on my
computer and I realized it was because of the default sleep of 100 ms.
To make this more flexible, we can add a sleep duration parameter which
can be set to zero in testing. The execution time for RetrySpec drops
from about 10 seconds to 10ms when turbo is enabled.