Avoid ConcurrentModificationException
I saw a ConcurrentModificationException in an sbt session while
traversing the observers map in Observers.onNext. This exception can
generally be avoided by using a ConcurrentHashMap rather than a
LinkedHashMap. In general, it shouldn't matter what order the
callbacks are evaluated so it feels ok to switch to the
ConcurrentHashMap even though we lose the insertion ordering.

In addition to changing observers to a ConcurrentHashMap, we can drop
the conversion of the observables map from a java.util.WeakHashMap to a
scala.collection.mutable.Map. There wasn't really any benefit to this
conversion since java 8 maps have the forEach method.