Remove millisecond-precision file timestamp using native code
**Problem**
Ref https://github.com/sbt/sbt/issues/7463
Ref https://github.com/sbt/sbt/issues/7455

On some Linux sbt 1.9.8 fail with "java.lang.UnsatisfiedLinkError: Error looking up function 'stat'".
This is due to our workaround another stat related-issue.
The general issue is that we currently call native call to retrieve
the last modified time because JDK 8 use to have a bug JDK-8177809
that failed to get milliseconds.
The bug was fixed in 2021 openjdk8u302.

**Solution**
We can drop the whole native code mechanism.